Apply some of the suggestions by Android Studio code inspection

+++ app/src/main/java/org/comixwall/pffw/Lib/org/comixwall/pffw/RefreshTimer.java
@@ -49,8 +49,8 @@ class RefreshTimer extends Timer {
         }
     };
 
-    public void start(int timeout) {
-        final int ONE_SECOND = 1000;
+    public void start(long timeout) {
+        final long ONE_SECOND = 1000;
         schedule(new RefreshTask(), timeout * ONE_SECOND, timeout * ONE_SECOND);
     }
 }
